[MIG] web_group_expand: Migration to 15.0
