--- web_group_expand/views/templates.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template
-        id="assets_backend"
-        name="web_groupby_expand assets"
-        inherit_id="web.assets_backend"
-    >
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/web_group_expand/static/src/js/web_group_expand.js"
-            />
-        </xpath>
-    </template>
-</odoo>
